Breaking L.Stats into own class, just in case
