+++ test/stat_tracker_test.rb
@@ -4,47 +4,3 @@ require_relative './stat_tracker'
 require_relative './games.csv'
 require_relative './teams.csv'
 require_relative './game_teams.csv'
-
-
-
-
-
-
-
-
-
-#LEAGUE STATS TESTS
-# 	Total number of teams in the data.  STRING INTEGER
-  def test_count_of_teams
-
-  end
-
-#  Name of the team with the highest average number of goals scored per game across all seasons.  STRING
-  def test_best_offense
-
-  end
-
- # Name of the team with the lowest average number of goals scored per game across all seasons.  STRING
-  def test_worst_offense
-
-  end
-
-# Name of the team with the highest average score per game across all seasons when they are away.  STRING
-  def test_highest_scoring_visitor
-
-  end
-
-# Name of the team with the highest average score per game across all seasons when they are home.  STRING
-  def test_highest_scoring_home_team
-
-  end
-
- # Name of the team with the lowest average score per game across all seasons when they are a visitor.  STRING
-  def test_lowest_scoring_visitor
-
-  end
-
- # 	Name of the team with the lowest average score per game across all seasons when they are at home.  STRING
-  def test_lowest_scoring_home_team
-
-  end
